Split cartrigde.c into cartrige.c and launcher.c
